==> 2_feature_binning.ipynb <==
### 1. Import Dependencies

How Binning Helps?

1. Non-linear relationship with target
    - Binning can help capture non-linear patterns that a linear model might miss

2. Skewed distribution
    - Binning can smooth out skew and reduce the effect of extreme values

3. Interpretability is key
    - Easier for business users to understand "age 18-25" than "age = 23"

4. Model is prone to overfitting (performing very well for traning data set, but not for test dataset)
    - Binning reduces granularity -> fewer splits -> less overfitting

5. Need to reduce cardinality
    - Helps when a numeric column has too many unique values

6. Sparse or noisy data
    - Binning can group rare or noisy values to improve signal strength

import os
import pandas as pd #alias
import numpy as np #alias
import seaborn as sns
from matplotlib import pyplot as plt

### 2. Basic Processing

df = pd.read_csv('processed/ChurnModelling_Outliers_Handled.csv')
df.head()

#### For this specific dataset we can apply binning like this (for numerical data),

- When you apply Binning you must know exactly the ranges and values. You cant create binning with abnormal intervals.
- if you dint have an clear visibility of what is the upper bound and what is the lower bound would be in the ranges 'Binning' would not be a ideal choice.
- In the binning the ranges dont need to be exactly similar distributed but cant have unrealistic ranges.

#### Eligible 
1. Age - we can apply binning to Age like mentioned below (but this is way too common)
2. EstiamtedSalary - also we can apply binning to this but what happens if a person comes up with a lot more money??
3. Credit Score - we exactly know 350 and 850 is a margin, so this works
4. Tenure - good variable to apply but there are some understanding that needs to be done in this

#### Not Eligible
1. Geograpgy - no point of adding binning because this is already is a categorical variable
2. Balance - we dont know the upper margin?? what is people come with 100M we cant make big bounds with abnormal margins.
3. Gender, NumOfProducts, HasCrCard, IsActiveMember - is already an categorical bin


df['Age'].min(), df['Age'].max()

for Age we can simply do like this as feature binning groups:

- 18 - 30 => Youngters
- 30 - 50 => Middle Age
- 50+ => Seniors

df['CreditScore'].min(), df['CreditScore'].max()

for CreditScore we can simply do like this as feature binning groups:

- 350 - 580 => 'Poor'
- 580 - 670 => 'Fair'
- 670 - 740 => 'Good'
- 740 - 800 => 'Very Good'
- 800 - 850 => 'Excellent'

def custom_binning_credit_score(score):
    if score < 580:
        return 'Poor'
    if score < 670:
        return 'Fair'
    if score < 740:
        return 'Good'
    if score < 800:
        return 'Very Good'
    if score <= 850:
        return 'Excellent'
    else:
        assert True, "Credit Score can't go beyond 850"
        
df['CreditScoreBins'] = df['CreditScore'].apply(custom_binning_credit_score)
del df['CreditScore']

df 

df.to_csv('processed/ChrunModelling_Binning_Applied.csv', index=False)
