==> 3_feature_encoding.ipynb <==
### 1. Import Dependencies

import os
import pandas as pd #alias
import numpy as np #alias
import seaborn as sns
from matplotlib import pyplot as plt

| Variable     |   Preferred Encoding   | Why?                                                                 |
|--------------|------------------------|----------------------------------------------------------------------|
| Nominal      | One-Hot Encoding       | No inherent order -> Avoids implying false ordinal relationships     |
| Ordinal      | Label Encoding         | Preserves order -> Small integers represent increasing levels        |

- Nominal : means that there are no order. (ex: Gender, Eye Color, Type of Car, Hair Color, Political Party)
- Ordinal : means that there are some kind of a order. (ex: Educational Level, Income Level, Customer Satisfaction, Age Range)

`Gender <Male, Female>`   // we will create 'n'  number of columns based on the number of diffrent columns.

|          |  Gender_Male | Gender_Female  |
|----------|--------------|----------------|
| Male     |       1      |       0        |
| Female   |       0      |       1        |

this is how the one-hot encoding representation usually works like.

`(Assume Gender is Ordinal);`
Male -> 1
Female -> 0


In this dataset these are the Values representations:

- Gender -> Nominal
- Geography -> Nominal
- CreditScoreBins -> Ordinal

df = pd.read_csv('processed\ChrunModelling_Binning_Applied.csv')
df.head()

### 2. Encode Nominal Variables

nominal_variables = ['Geography', 'Gender']   #putting them in a list

# Create dummy variables
geography_dummies = pd.get_dummies(df['Geography'], prefix="Geography")
gender_dummies = pd.get_dummies(df['Gender'], prefix="Gender")

# Start with original df
df_encoded = df.copy()

# Add dummy variables
df_encoded = pd.concat([df_encoded, geography_dummies, gender_dummies], axis=1)

# Drop original categorical columns
df_encoded.drop(['Geography', 'Gender'], axis=1, inplace=True)

df_encoded


### 3. Encode Ordinal Variables

encode_dict_creditscore = {
                                'Poor'      : 0,
                                'Fair'      : 1,
                                'Good'      : 2,
                                'Very Good' : 3,
                                'Excellent' : 4
                            }

df_encoded['CreditScoreBins'] = df_encoded['CreditScoreBins'].map(encode_dict_creditscore)
df_encoded

- Now all of them are in numerical formats, but in different ranges.
- We need to fix that.

df_encoded.to_csv('processed/ChurnModelling_Encoded.csv', index=False)

### Also you can do these with Scikit-Learn as well.

from sklearn.preprocessing import LabelEncoder, OneHotEncoder

df = pd.read_csv('processed\ChrunModelling_Binning_Applied.csv')
df

ohe_geography = OneHotEncoder()
ohe_gender = OneHotEncoder()

le_credit_score = LabelEncoder()

ohe_geography.fit(df['Geography'].values.reshape(10000, 1))
ohe_gender.fit(df['Gender'].values.reshape(10000, 1))

geography_ohe = ohe_geography.transform(df['Geography'].values.reshape(10000, 1))
geography_ohe = geography_ohe.toarray()
geography_ohe

gender_ohe = ohe_gender.transform(df['Gender'].values.reshape(10000, 1))
gender_ohe = gender_ohe.toarray()
gender_ohe
