==> 4_feature_scaling.ipynb <==
### 1. Import Dependencies

import os
import pandas as pd #alias
import numpy as np #alias
import seaborn as sns
from matplotlib import pyplot as plt
from sklearn.preprocessing import StandardScaler, MinMaxScaler

### 2. Important Concepts

### 2.1 Normalization vs Standardization

#### 2.1.1 What is Normalization?
    - Normalization is a scalling technique in which values are shifted and rescaled so that they end up ranging between 0 and 1. it is also known as `Min-Max Scalling`

#### 2.1.2 What is Standardization?
    - Standardization is another scalling technique where the values are centered around the mean with a unit standard deviation. This means that the mean of the attribute beacomes zero(0) and the resultant distribution has a unit standard deviation. 

mean = 0

0 - 1, 0 + 1 => -1, 1

df = pd.read_csv('processed\ChurnModelling_Encoded.csv')
df.head(10)

| Condition                                        | Min-Max Scaling                      | Standardization(Z-Score)                  |
|--------------------------------------------------|--------------------------------------|-------------------------------------------|
| Data has a known, fixed range                    | ✅ Yes                               | ❌ Not ideal                             |                                 
| Data contains outliers                           | ❌ Sensitive to outliers             | ✅ More robust to outliers               |                                                        
| Data is normally distributed                     | ❌ Not necessary                     | ✅ Preferred                             |
| Data is not normally distributed (eg: skewed..)  | ✅ If shape needs to be preserved    | ✅ Often works well after log-transform  |
| Model is distance-based (KNN, SVM)               | ✅ Recommended                       | ✅ Also acceptable                       |                                                                
| Model is neural network                          | ✅ Strongly recommended              | ❌ May slow training                     |                                                       
| Model is linear or uses regularization           | ❌ Not ideal                         | ✅ Helps with covergence                 |                                              
| Input features need bounded values (0-1)         | ✅ Required                          | ❌ Not bounded                           |                              
| Applying PCA or LDA                              | ❌ May distort variance              | ✅ Required (centering needed)           |                                                       
| Want to preserve original distribution shape     | ✅ Maintains feature shape           | ✅ Maintains shape but centers data      |                                         
| Working with tree-based models                   | ❌ Not needed                        | ❌ Not needed                            |            

columns_need_to_be_scaled = ['Age', 'Tenure', 'Balance', 'EstimatedSalary']

for col in columns_need_to_be_scaled:
    scaler = StandardScaler()
    # scaler = MinMaxScaler()   # you can see there are no negative values when Min-Max used
    df[col] = scaler.fit_transform(df[col].values.reshape(10000, 1))
    
df

df.to_csv('processed/ChurnModelling_Final.csv', index=False)

### This would be the final dataset that we will be ready to feed into the models.
